Replace existing dashboard with Morning Dashboard for Dr. Gatsas - Replace DentalSleepDashboard with Morning Dashboard implementation - Remove separate Morning Dashboard tab and route - Update navigation to use single dashboard - Implement all PRD requirements in main dashboard - Add Today's Patient Schedule with status indicators - Add Yesterday's Incomplete Cases section - Add Action Items Queue with priority levels - Add Quick Stats overview with key metrics - Include call/message/view functionality for patients - Ready for production deployment
