+++ docs/team/DEVELOPMENT_STATUS.md
@@ -33,8 +33,8 @@
 1. **Platform Admin Dashboard** (`src/components/admin/PlatformTenants.tsx`)
    - Status: In Progress
    - Branch: `production-deployment-ready`
-   - Last Commit: `8dea138 - Add debugging to edit appointment functionality - Add console.log to handleEditAppointment function - Add console.log to Edit modal render - Debug why edit button is not working`
-   - Uncommitted Changes:       89 files
+   - Last Commit: `9af6111 - Implement complete Morning Dashboard for Dr. Gatsas - Add MorningDashboard component with all PRD requirements - Remove debugging console.log statements from Schedule.tsx - Add Morning Dashboard to navigation and routing - Include Today's Patient Schedule with status indicators - Add Yesterday's Incomplete Cases section - Implement Action Items Queue with priority levels - Add Quick Stats overview with key metrics - Include call/message/view functionality for patients - Add proper TypeScript interfaces and mock data - Ready for production deployment`
+   - Uncommitted Changes:        5 files
    - Next: Complete tenant management features
 
 2. **Production Deployment Verification**
@@ -144,5 +144,5 @@ Next: Practice setup and user onboarding
 
 ---
 
-**Last Updated**: August 05, 2025 02:16 PM CT  
+**Last Updated**: August 05, 2025 02:24 PM CT  
 **Next Update**: Next development session
