+++ docs/team/TEAM_SPRINT_BOARD.md
@@ -443,3 +443,17 @@
 
 **Last Updated**: August 05, 2025 02:16 PM CT  
 **Next Standup**: August 06, 2025 10:00 AM CT
+
+### **August 05, 2025 02:24 PM CT - Jason**
+- ✅ Created comprehensive team coordination system
+- ✅ Verified git workflow and CI/CD pipeline
+- 🔄 Next: Complete platform admin dashboard features
+
+### **August 05, 2025 02:24 PM CT - Jeff**
+- ✅ Team coordination system ready for use
+- 🔄 Next: Begin practice setup configuration
+
+---
+
+**Last Updated**: August 05, 2025 02:24 PM CT  
+**Next Standup**: August 06, 2025 10:00 AM CT
