+++ docs/team/THREAD_CONTINUITY.md
@@ -49,7 +49,7 @@ What specific task should I help with today?"
 - **CI/CD Pipeline**: ✅ Fully operational
 - **Branch Strategy**: main ← develop ← feature/[developer]/[feature]
 - **Current Branch**: production-deployment-ready
-- **Last Commit**: 8dea138 - Add debugging to edit appointment functionality - Add console.log to handleEditAppointment function - Add console.log to Edit modal render - Debug why edit button is not working
+- **Last Commit**: 9af6111 - Implement complete Morning Dashboard for Dr. Gatsas - Add MorningDashboard component with all PRD requirements - Remove debugging console.log statements from Schedule.tsx - Add Morning Dashboard to navigation and routing - Include Today's Patient Schedule with status indicators - Add Yesterday's Incomplete Cases section - Implement Action Items Queue with priority levels - Add Quick Stats overview with key metrics - Include call/message/view functionality for patients - Add proper TypeScript interfaces and mock data - Ready for production deployment
 
 ---
 
@@ -126,5 +126,5 @@ What specific task should I help with today?"
 
 ---
 
-**Last Updated**: August 05, 2025 01:58 PM CT  
+**Last Updated**: August 05, 2025 02:16 PM CT  
 **Next Review**: Next development session
